+++ tests/api/utils/test_naming_utils.py
@@ -19,7 +19,7 @@ from snowflake.cli.api.identifiers import FQN
 @pytest.mark.parametrize(
     "qualified_name, expected",
     [
-        ("func(number, number)", ("func(number, number)", None, None)),
+        ("func(number, number)", ("func", None, None)),
         ("name", ("name", None, None)),
         ("schema.name", ("name", "schema", None)),
         ("db.schema.name", ("name", "schema", "db")),
@@ -31,3 +31,5 @@ def test_from_fully_qualified_name(qualified_name, expected):
     assert fqn.name == name
     assert fqn.schema == schema
     assert fqn.database == database
+    if fqn.signature:
+        assert fqn.signature == "(number, number)"