+++ RELEASE-NOTES.md
@@ -24,6 +24,20 @@
 ## Fixes and improvements
 
 
+# v2.8.1
+## Backward incompatibility
+
+## Deprecations
+
+## New additions
+
+## Fixes and improvements
+* Fixed git execute not working with upper case in directory name.
+* Fixed `snow git setup` command behaviour for fully qualified repository names.
+* Fixed `snow git setup` command behaviour in case API integration or secret with default name already exists.
+* Fixed `snow snowpark package create` creating empty zip when package name contained capital letters.
+
+
 # v2.8.0
 ## Backward incompatibility